Add files via upload
First, a screen appears to check if you are the correct player. Then, each player seens a screen with everyone's guesses, wins, picked cards, the trump card, and that player's hand. The player then clicks on one of the radio button cards and that card is now in the player's picked variable and pop out of its hand.